refactor: storeService와 purchaseService의 역할을 명확히 분리함
store는 일반적인 조회와 영수증을 처리, purchase는 결제 관련 내용을
처리함
