+++ src/main/java/store/service/PurchaseService.java
@@ -1,6 +1,7 @@
 package store.service;
 
 import static store.constants.StoreErrorCode.CANNOT_ANSWER_PROCESS;
+import static store.constants.StoreErrorCode.CANNOT_REQUEST_OVER_STORED_QUANTITY;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,9 +14,6 @@ import promotion.service.PromotionService;
 import store.dto.request.PurchaseForm;
 import store.dto.request.PurchaseRequest;
 import store.dto.request.SeparatedPurchaseRequest;
-import store.dto.FinalBonus;
-import store.dto.FinalPurchase;
-import store.dto.response.ReceiptItems;
 
 public class PurchaseService {
 
@@ -27,11 +25,20 @@ public class PurchaseService {
         this.promotionService = promotionService;
     }
 
+    public void checkPurchaseRequests(List<PurchaseRequest> purchaseRequests) {
+        for (PurchaseRequest request : purchaseRequests) {
+            productService.getByName(request.productName());
+            int storedQuantity = productService.getAllQuantityByName(request.productName());
+            if (request.quantity() > storedQuantity) {
+                throw new IllegalArgumentException(CANNOT_REQUEST_OVER_STORED_QUANTITY.getMsgWithPrefix());
+            }
+        }
+    }
+
     public SeparatedPurchaseRequest separateRequest(List<PurchaseRequest> requests) {
         List<Promotion> activePromotions = promotionService.getActivePromotions();
         List<PurchaseRequest> promotionRequests = new ArrayList<>();
         List<PurchaseRequest> normalRequests = new ArrayList<>();
-
         for (PurchaseRequest request : requests) {
             if (productService.isPromotionTargetRequest(activePromotions, request)) {
                 promotionRequests.add(request);
@@ -39,19 +46,15 @@ public class PurchaseService {
             }
             normalRequests.add(request);
         }
-
         return new SeparatedPurchaseRequest(promotionRequests, normalRequests);
     }
 
     public List<PromotionApplyResult> promotionApplyRequest(List<PurchaseRequest> promotionRequests) {
-
         List<PromotionApplyResult> results = new ArrayList<>();
-
         for (PurchaseRequest request : promotionRequests) {
             Product product = productService.getByNameAndHasPromotion(request.productName());
             results.add(promotionService.getPromotionApplyResult(product, request.quantity()));
         }
-
         return results;
     }
 
@@ -59,11 +62,9 @@ public class PurchaseService {
         if (result.state() == PromotionApplyState.ADDITIONAL_PROMOTION_AVAILABLE) {
             return fixForAdditionalCondition(isPositiveAnswer, result);
         }
-
         if (result.state() == PromotionApplyState.PARTIAL_PROMOTION_APPLIED) {
             return fixForPartialCondition(isPositiveAnswer, result);
         }
-
         throw new IllegalArgumentException(CANNOT_ANSWER_PROCESS.getMsgWithPrefix());
     }
 
@@ -84,10 +85,8 @@ public class PurchaseService {
     public List<PurchaseForm> processingPurChaseRequests(List<PromotionApplyResult> promotionResults,
                                                          List<PurchaseRequest> normalRequests) {
         List<PurchaseForm> purchaseForms = new ArrayList<>();
-
         purchaseForms.addAll(getPurchaseFormByPromotion(promotionResults));
         purchaseForms.addAll(getPurchaseFormByNormal(normalRequests));
-
         return purchaseForms;
     }
 
@@ -107,12 +106,10 @@ public class PurchaseService {
 
     private List<PurchaseForm> separatedPurchaseFormForPromotion(Product product, int requiredQuantity) {
         List<PurchaseForm> purchaseForms = new ArrayList<>();
-
         Product nonePromotionProduct = productService.getByNameAndNotHasPromotion(product.getName());
         purchaseForms.add(new PurchaseForm(product, product.getQuantity()));
         purchaseForms.add(new PurchaseForm(nonePromotionProduct
                 , calcRemainingQuantity(requiredQuantity, product.getQuantity())));
-
         return purchaseForms;
     }
 
@@ -134,60 +131,17 @@ public class PurchaseService {
         List<PurchaseForm> purchaseForms = new ArrayList<>();
         int normalQuantity = normalProduct.getQuantity();
         Product promotionProduct = productService.getByNameAndHasPromotion(normalProduct.getName());
-
         purchaseForms.add(new PurchaseForm(normalProduct, normalProduct.getQuantity()));
         purchaseForms.add(new PurchaseForm(promotionProduct,
                 calcRemainingQuantity(requiredQuantity, normalQuantity)));
-
         return purchaseForms;
-    }
-
-    private int calcRemainingQuantity(int requiredQuantity, int quantityForPromotion) {
-        return requiredQuantity - quantityForPromotion;
     }
 
     public void purchase(List<PurchaseForm> purchaseForms) {
         productService.purchase(purchaseForms);
     }
 
-    public ReceiptItems processingReceiptItems(List<PromotionApplyResult> promotionResults,
-                                               List<PurchaseRequest> normalRequests) {
-        ReceiptItems promotionReceiptItems = convertToReceiptItemByPromotion(promotionResults);
-        ReceiptItems normalReceiptItems = convertToReceiptItemByNormal(normalRequests);
-
-        return integrateReceiptItems(promotionReceiptItems, normalReceiptItems);
-    }
-
-    public ReceiptItems convertToReceiptItemByPromotion(List<PromotionApplyResult> promotionResults) {
-        List<FinalPurchase> finalPurchases = new ArrayList<>();
-        List<FinalBonus> finalBonuses = new ArrayList<>();
-        for (PromotionApplyResult result : promotionResults) {
-            Product product = result.product();
-            int quantity = result.promotionPurchase();
-            int bonusQuantity = result.bonusQuantity();
-            int price = product.getPrice();
-            finalPurchases.add(new FinalPurchase(product.getName(), quantity, price * quantity));
-            finalBonuses.add(new FinalBonus(product.getName(), bonusQuantity, price * bonusQuantity));
-        }
-        return new ReceiptItems(finalPurchases, finalBonuses);
-    }
-
-    public ReceiptItems convertToReceiptItemByNormal(List<PurchaseRequest> normalRequests) {
-        List<FinalPurchase> finalPurchases = new ArrayList<>();
-        for (PurchaseRequest request : normalRequests) {
-            String productName = request.productName();
-            Product product = productService.getByNameAndNotHasPromotion(productName);
-            finalPurchases.add(new FinalPurchase(request.productName(), request.quantity()
-                    , product.getPrice() * request.quantity()));
-        }
-        return new ReceiptItems(finalPurchases, null);
-    }
-
-    private ReceiptItems integrateReceiptItems(ReceiptItems promotionReceiptItems, ReceiptItems normalReceiptItems) {
-        List<FinalPurchase> integratedPurchases = new ArrayList<>();
-        integratedPurchases.addAll(promotionReceiptItems.purchases());
-        integratedPurchases.addAll(normalReceiptItems.purchases());
-
-        return new ReceiptItems(integratedPurchases, promotionReceiptItems.bonuses());
+    private int calcRemainingQuantity(int requiredQuantity, int quantityForPromotion) {
+        return requiredQuantity - quantityForPromotion;
     }
 }
