+++ src/main/java/store/service/StoreService.java
@@ -2,11 +2,13 @@ package store.service;
 
 import static product.constants.ProductStatic.PRODUCT_FILE_PATH;
 import static promotion.constants.PromotionStatic.PROMOTION_FILE_PATH;
-import static store.constants.StoreErrorCode.CANNOT_REQUEST_OVER_STORED_QUANTITY;
 
+import java.util.ArrayList;
 import java.util.List;
 import product.dto.ProductInfo;
+import product.entity.Product;
 import product.service.ProductService;
+import promotion.dto.response.PromotionApplyResult;
 import promotion.service.PromotionService;
 import store.dto.request.PurchaseRequest;
 import store.dto.FinalBonus;
@@ -35,14 +37,45 @@ public class StoreService {
         return productService.getProductInfos();
     }
 
-    public void checkPurchaseRequests(List<PurchaseRequest> purchaseRequests) {
-        for (PurchaseRequest request : purchaseRequests) {
-            productService.getByName(request.productName());
-            int storedQuantity = productService.getAllQuantityByName(request.productName());
-            if (request.quantity() > storedQuantity) {
-                throw new IllegalArgumentException(CANNOT_REQUEST_OVER_STORED_QUANTITY.getMsgWithPrefix());
-            }
+    public ReceiptItems processingReceiptItems(List<PromotionApplyResult> promotionResults,
+                                               List<PurchaseRequest> normalRequests) {
+        ReceiptItems promotionReceiptItems = convertToReceiptItemByPromotion(promotionResults);
+        ReceiptItems normalReceiptItems = convertToReceiptItemByNormal(normalRequests);
+
+        return integrateReceiptItems(promotionReceiptItems, normalReceiptItems);
+    }
+
+    public ReceiptItems convertToReceiptItemByPromotion(List<PromotionApplyResult> promotionResults) {
+        List<FinalPurchase> finalPurchases = new ArrayList<>();
+        List<FinalBonus> finalBonuses = new ArrayList<>();
+        for (PromotionApplyResult result : promotionResults) {
+            Product product = result.product();
+            int quantity = result.promotionPurchase();
+            int bonusQuantity = result.bonusQuantity();
+            int price = product.getPrice();
+            finalPurchases.add(new FinalPurchase(product.getName(), quantity, price * quantity));
+            finalBonuses.add(new FinalBonus(product.getName(), bonusQuantity, price * bonusQuantity));
         }
+        return new ReceiptItems(finalPurchases, finalBonuses);
+    }
+
+    public ReceiptItems convertToReceiptItemByNormal(List<PurchaseRequest> normalRequests) {
+        List<FinalPurchase> finalPurchases = new ArrayList<>();
+        for (PurchaseRequest request : normalRequests) {
+            String productName = request.productName();
+            Product product = productService.getByNameAndNotHasPromotion(productName);
+            finalPurchases.add(new FinalPurchase(request.productName(), request.quantity()
+                    , product.getPrice() * request.quantity()));
+        }
+        return new ReceiptItems(finalPurchases, null);
+    }
+
+    private ReceiptItems integrateReceiptItems(ReceiptItems promotionReceiptItems, ReceiptItems normalReceiptItems) {
+        List<FinalPurchase> integratedPurchases = new ArrayList<>();
+        integratedPurchases.addAll(promotionReceiptItems.purchases());
+        integratedPurchases.addAll(normalReceiptItems.purchases());
+
+        return new ReceiptItems(integratedPurchases, promotionReceiptItems.bonuses());
     }
 
     public ReceiptPriceInfo processingReceiptPriceInfo(ReceiptItems items, boolean isContainsMembershipDiscount) {
